source/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


# need either GhostScript ot pdftoppm installed and in path
numpy         
opencv-python 
Pillow    
pdfCropMargins   
imutils